+++ src/models/catalog.ts
@@ -8,7 +8,7 @@ export default class Catalog extends ArrayLike<string> {
   // From https://scryfall.com/docs/api/catalogs
   // omitting data since it becomes an entry in the array like structure
   object: "catalog";
-  uri: URL; // A link to the current catalog on Scryfall’s API.
+  uri: string; // A link to the current catalog on Scryfall’s API.
   total_values: number; // The number of items in the data array.
 
   constructor(scrfallResponse: CatalogApiResponse) {