+++ test/unit/models/card.test.ts
@@ -8,15 +8,13 @@ import { get } from "Lib/api-request";
 import wrapScryfallResponse from "Lib/wrap-scryfall-response";
 import fixtures from "Fixtures";
 
-import { mocked } from "ts-jest/utils";
-
 jest.mock("Lib/api-request");
 
 describe("Card", function () {
   let fakeRequest: jest.SpyInstance, card: Card;
 
   beforeEach(() => {
-    fakeRequest = mocked(get);
+    fakeRequest = jest.mocked(get);
   });
 
   afterEach(() => {